Removing code from index.js to firebase.js
